Rename `Reckoning` sub-decision to `Statement`
`reckoning` is a grossly inaccurate translation of `afrekening` in
the context of finances. With this change we now have a financial
`budget` and `statement`.

These sub-decisions have also been moved into their own category.
